test/requirements.txt
pyserial
keyboard
flask
flask_cors